requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydub==0.25.1
requests==2.31.0
oss2==2.18.4
aiofiles==23.2.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0
